feat: Add .ssignore support
Allow each scaffold to contain .ssignore file which contains a collection of files and folders
to be excluded from processing during the provisioning.
The concept of .ssignore file is similar to of .gitignore, although only basic wildcards are currently supported.

If a scaffold does not contain own copy of .ssignore file then the default set of ignores is applied to the
scaffold (".git", "node_modules", "bower_components", "packages", "bin", "obj")